comments + move constants to separate directory
