+++ packages/cloudflare/src/cli/templates/cache-handler/index.ts
@@ -1,2 +1,1 @@
-export * from "./constants";
 export * from "./open-next-cache-handler";
