+++ packages/cloudflare/src/cli/templates/cache-handler/utils.ts
@@ -1,5 +1,5 @@
 import type { IncrementalCache } from "next/dist/server/lib/incremental-cache";
-import { NEXT_META_SUFFIX } from "./constants";
+import { NEXT_META_SUFFIX } from "../../constants/incremental-cache";
 
 type PrerenderedRouteMeta = {
   lastModified: number;
